refactor: move the seismic-intensity module into shared/seismic
