+++ lib/features/earthquake/domain/eew_local_estimate.dart
@@ -1,7 +1,7 @@
 import 'package:dpip/core/models/lat_lng.dart';
 import 'package:dpip/features/earthquake/domain/eew.dart';
 import 'package:dpip/features/earthquake/domain/eew_estimator.dart';
-import 'package:dpip/features/earthquake/domain/intensity.dart';
+import 'package:dpip/shared/seismic/intensity.dart';
 import 'package:dpip/features/earthquake/domain/seismic_travel_time.dart';
 
 /// The EEW shaking estimate for one observer location.
