+++ lib/shared/seismic/intensity.dart
